requirements.txt
alembic==1.13.2
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dropbox==12.0.2
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
ply==3.11
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
stone==3.3.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
